backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai-whisper==20231117
anthropic==0.7.7
openai==1.3.7
python-dotenv==1.0.0
aiofiles==23.2.1